New mechanic for Sayoristein
- Added the "Sayo-Forge":
It includes a space to upgrade things only for Arena Mode.
When you kill an enemy, they have 35% prob to drop coins, each coins has the rate of 100 to 500 to give you coins, in bosses the prob is 100% but with a base bonus of 200% of coins.
- Availables upgrade:
Level up gun damage
Level up shotgun damage